requirements.txt
playwright>=1.40.0
openai>=1.3.0
python-dotenv>=1.0.0
flask>=2.3.3
flask-cors>=4.0.0
requests>=2.28.0

